bank/tx: make postings use bigdec

+++ bank-service/src/main/java/rs/banka4/bank_service/tx/data/Posting.java
@@ -1,8 +1,10 @@
 package rs.banka4.bank_service.tx.data;
 
+import java.math.BigDecimal;
+
 public record Posting(
     TxAccount account,
-    int amount,
+    BigDecimal amount,
     TxAsset asset
 ) {
 }
